app/requirements.txt
requests == 2.22.0
python-telegram-bot == 12.2.0
Flask == 1.1.1